requirements.txt
streamlit
supabase
python-dotenv
requests
deepgram-sdk
python-vlc
audioop-lts
PyObjC